rename views to treeviews
