+++ src/commands/commands.ts
@@ -1,35 +1,35 @@
 import { commands, Disposable, ExtensionContext, Uri, window } from 'vscode';
-import { copyResourceName } from './commands/copyResourceName';
-import { createGitRepositoryForPath } from './commands/createGitRepositoryForPath';
-import { createKustomizationForPath } from './commands/createKustomizationForPath';
-import { addSource } from './commands/addSource';
-import { addKustomization } from './commands/addKustomization';
-import { deleteWorkload } from './commands/deleteWorkload';
-import { deleteSource } from './commands/deleteSource';
-import { fluxDisableGitOps, fluxEnableGitOps } from './commands/enableDisableGitOps';
-import { fluxCheck } from './commands/fluxCheck';
-import { checkFluxPrerequisites } from './commands/fluxCheckPrerequisites';
-import { fluxReconcileRepositoryForPath } from './commands/fluxReconcileGitRepositoryForPath';
-import { fluxReconcileSourceCommand } from './commands/fluxReconcileSource';
-import { fluxReconcileWorkload } from './commands/fluxReconcileWorkload';
-import { installFluxCli } from './commands/installFluxCli';
-import { openResource } from './commands/openResource';
-import { pullGitRepository } from './commands/pullGitRepository';
-import { resume } from './commands/resume';
-import { setClusterProvider } from './commands/setClusterProvider';
-import { setCurrentKubernetesContext } from './commands/setCurrentKubernetesContext';
-import { showGlobalState } from './commands/showGlobalState';
-import { showInstalledVersions } from './commands/showInstalledVersions';
-import { showNewUserGuide } from './commands/showNewUserGuide';
-import { showLogs } from './commands/showLogs';
-import { showWorkloadsHelpMessage } from './commands/showWorkloadsHelpMessage';
-import { suspend } from './commands/suspend';
-import { trace } from './commands/trace';
-import { telemetry } from './extension';
-import { showOutputChannel } from './output';
-import { TelemetryErrorEventNames } from './telemetry';
-import { refreshAllTreeViews, refreshResourcesTreeViews } from './views/treeViews';
-import { createFromTemplate } from './commands/createFromTemplate';
+import { copyResourceName } from './copyResourceName';
+import { createGitRepositoryForPath } from './createGitRepositoryForPath';
+import { createKustomizationForPath } from './createKustomizationForPath';
+import { addSource } from './addSource';
+import { addKustomization } from './addKustomization';
+import { deleteWorkload } from './deleteWorkload';
+import { deleteSource } from './deleteSource';
+import { fluxDisableGitOps, fluxEnableGitOps } from './enableDisableGitOps';
+import { fluxCheck } from './fluxCheck';
+import { checkFluxPrerequisites } from './fluxCheckPrerequisites';
+import { fluxReconcileRepositoryForPath } from './fluxReconcileGitRepositoryForPath';
+import { fluxReconcileSourceCommand } from './fluxReconcileSource';
+import { fluxReconcileWorkload } from './fluxReconcileWorkload';
+import { installFluxCli } from './installFluxCli';
+import { openResource } from './openResource';
+import { pullGitRepository } from './pullGitRepository';
+import { resume } from './resume';
+import { setClusterProvider } from './setClusterProvider';
+import { setCurrentKubernetesContext } from './setCurrentKubernetesContext';
+import { showGlobalState } from './showGlobalState';
+import { showInstalledVersions } from './showInstalledVersions';
+import { showNewUserGuide } from './showNewUserGuide';
+import { showLogs } from './showLogs';
+import { showWorkloadsHelpMessage } from './showWorkloadsHelpMessage';
+import { suspend } from './suspend';
+import { trace } from './trace';
+import { telemetry } from '../extension';
+import { showOutputChannel } from '../output';
+import { TelemetryErrorEventNames } from '../telemetry';
+import { refreshAllTreeViews, refreshResourcesTreeViews } from '../treeviews/treeViews';
+import { createFromTemplate } from './createFromTemplate';
 
 /**
  * Command ids registered by this extension
