+++ src/commands/createKustomizationForPath.ts
@@ -3,7 +3,7 @@ import { Uri, window, workspace } from 'vscode';
 import { failed } from '../errorable';
 import { getFolderGitInfo, getGitRepositoryforGitInfo } from '../git/gitInfo';
 import { namespacedObject } from '../kubernetes/types/flux/object';
-import { getCurrentClusterInfo } from '../views/treeViews';
+import { getCurrentClusterInfo } from '../treeviews/treeViews';
 import { openConfigureGitOpsWebview } from '../webview-backend/configureGitOps/openWebview';
 
 /**
