+++ src/commands/deleteSource.ts
@@ -6,11 +6,11 @@ import { fluxTools } from '../flux/fluxTools';
 import { FluxSource } from '../flux/fluxTypes';
 import { KubernetesObjectKinds } from '../kubernetes/types/kubernetesTypes';
 import { TelemetryEventNames } from '../telemetry';
-import { BucketNode } from '../views/nodes/bucketNode';
-import { GitRepositoryNode } from '../views/nodes/gitRepositoryNode';
-import { OCIRepositoryNode } from '../views/nodes/ociRepositoryNode';
-import { HelmRepositoryNode } from '../views/nodes/helmRepositoryNode';
-import { getCurrentClusterInfo, refreshSourcesTreeView, refreshWorkloadsTreeView } from '../views/treeViews';
+import { BucketNode } from '../treeviews/nodes/bucketNode';
+import { GitRepositoryNode } from '../treeviews/nodes/gitRepositoryNode';
+import { OCIRepositoryNode } from '../treeviews/nodes/ociRepositoryNode';
+import { HelmRepositoryNode } from '../treeviews/nodes/helmRepositoryNode';
+import { getCurrentClusterInfo, refreshSourcesTreeView, refreshWorkloadsTreeView } from '../treeviews/treeViews';
 
 /**
  * Delete a source
