+++ src/test/suite/extension.test.ts
@@ -1,11 +1,11 @@
 import * as assert from 'assert';
 import { before } from 'mocha';
 import * as vscode from 'vscode';
-import { DataProvider } from '../../views/dataProviders/dataProvider';
-import { ClusterDataProvider } from '../../views/dataProviders/clusterDataProvider';
-import { SourceDataProvider } from '../../views/dataProviders/sourceDataProvider';
-import { WorkloadDataProvider } from '../../views/dataProviders/workloadDataProvider';
-import { NamespaceNode } from '../../views/nodes/namespaceNode';
+import { DataProvider } from '../../treeviews/dataProviders/dataProvider';
+import { ClusterDataProvider } from '../../treeviews/dataProviders/clusterDataProvider';
+import { SourceDataProvider } from '../../treeviews/dataProviders/sourceDataProvider';
+import { WorkloadDataProvider } from '../../treeviews/dataProviders/workloadDataProvider';
+import { NamespaceNode } from '../../treeviews/nodes/namespaceNode';
 
 let api: {
 	shell: { execWithOutput: any; };
