+++ src/treeviews/nodes/gitOpsTemplateNode.ts
