+++ src/treeviews/nodes/namespaceNode.ts
